fix: Handling of missing multiqc module
This commit contains the follwong updates:
- When running a simple pipeline as described in
https://hydra-genetics.readthedocs.io/en/latest/tutorials/simple_pipeline/
the workflow will fail due to a key error in the config.yaml file.
MultiQC is expected to always be included as a module. However, in this
simple case this key is missing in the configuration file. This will now
be handled and the workflow will not fail if multiqc is missing from the
configuration file. Instead a warning will be logged.
- Corrected imports in common.smk to correspond to function names in
  software_versions.py