fix: handle missing VEP annotation
When VEP > 109 can't annotate a variant it leaves the variant in the vcf file without a CSQ file. Previously the variant was filtered. This PR fixes this and returns None for all filter values. The individual filters NA handling will then determine if the variant should be filtered or not. The default is NA_FALSE which does not filter the variant.